IPC implementation in bootloader
+ Create the same IPC structure for bootloader.
+ Update callback function to have two char arrays as arguments (in and out).

However, this approach does not work after all...

The problem is that when we want to flash the crownstone binary or the
bootloader binary the code that is occupying page 0x6D000 will be
overwritten. There is always an erase, either by nrfjprog or the JLink
utility.

This means we will have to resort to a runtime implementation of https://github.com/crownstone/bluenet/issues/91